+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
  * added polyfill for `hrtime()`
  * added polyfills for `array_key_first()` and `array_key_last()`
+ * fixed infinite loop in `iconv()` polyfill when using translit mode
 
 * v1.8.0